Two Passes, SSAO, Fog
* Running 2 passes to capture depth
* Normals calculated using difference and then projected into the world
* SSAO
* Fog - based on depth and height
* Better tessellation based on distances even when rotating camera
* 2 levels of tessellation
* Added options to display each available mode

+++ source/GPUTerrainGenerator/GPUTerrainGenerator/Mesh.h
@@ -8,11 +8,11 @@ using namespace std;
 
 struct Triangle
 {
-	unsigned short x;
-	unsigned short y;
-	unsigned short z;
+	unsigned int x;
+	unsigned int y;
+	unsigned int z;
 
-	Triangle(unsigned short i_x, unsigned short i_y, unsigned short i_z)
+	Triangle(unsigned int i_x, unsigned int i_y, unsigned int i_z)
 	{
 		x = i_x;
 		y = i_y;
